requirements.txt
aiohttp
discord.py
ruamel.yaml
